fix errors, ready to push
